Line component and per-page dimensions
Adds a line component that has a hover area around it. Responds to
scaling to keep the line thickness.
Fixes a bug where the annotations overlay does not have the exact page's
width and height, but wrongly uses the width/ height of the "rescaled"
component which uses the max page size. This leads to wrong coordinates
of points